+++ src/bda/plone/orders/interfaces.py
@@ -30,28 +30,28 @@ class IPaymentText(Interface):
 
 
 class IDynamicMailTemplateLibrary(Interface):
-    """A set of named templates
+    """A set of named templates.
     """
 
     def keys():
-        """list names of templates
+        """list names of templates.
         """
 
     def __getitem__(name):
-        """return template by name
+        """return template by name.
         """
 
 
 class IDynamicMailTemplateLibraryStorage(IDynamicMailTemplateLibrary):
     
-    def direct_keys(self):
-        """non acquired keys
+    def direct_keys():
+        """non acquired keys.
         """
 
     def __setitem__(name, template):
-        """store template under a name
+        """store template under a name.
         """
 
     def __delitem__(name):
-        """remove template with this name
+        """remove template with this name.
         """