+++ src/bda/plone/orders/mailnotify.py
@@ -21,6 +21,8 @@ def status_message(context, msg):
 
 
 def create_mail_listing(context, attrs):
+    """Create item listing for notification mail.
+    """
     soup = get_soup('bda_plone_orders_bookings', context)
     bookings = soup.query((Any('uid', attrs['booking_uids'])))
     lines = []
@@ -36,6 +38,11 @@ def create_mail_listing(context, attrs):
 
 
 def create_order_total(context, attrs):
+    """Calculate order total price for notification mail.
+
+    XXX: use CartItemCalculator? Problem - lives in bda.plone.shop
+    XXX: also consider discount here once implemented
+    """
     soup = get_soup('bda_plone_orders_bookings', context)
     bookings = soup.query((Any('uid', attrs['booking_uids'])))
     ret = 0.0
@@ -45,6 +52,18 @@ def create_order_total(context, attrs):
         ret += net
         ret += net * booking.attrs.get('vat', 0.0) / 100
     return "%.2f" % (ret + float(attrs['shipping']))
+
+
+def create_order_text(context, attrs):
+    pass
+
+
+def create_reservation_text(context, attrs):
+    pass
+
+
+def create_payment_text(context, attrs):
+    pass
 
 
 def create_mail_body(templates, context, attrs):
@@ -90,6 +109,8 @@ def notify_payment_success(event):
     templates.update(get_order_templates(event.context))
     templates['item_listing_callback'] = create_mail_listing
     templates['order_total_callback'] = create_order_total
+    templates['order_text_callback'] = create_order_text
+    templates['payment_text_callback'] = create_payment_text
     do_notify(event.context, 'order', order, templates)
 
 
@@ -106,6 +127,8 @@ def notify_reservation_if_payment_skipped(event):
     templates.update(get_reservation_templates(event.context))
     templates['item_listing_callback'] = create_mail_listing
     templates['order_total_callback'] = create_order_total
+    templates['order_text_callback'] = create_reservation_text
+    templates['payment_text_callback'] = create_payment_text
     do_notify(event.context, 'reservation', order, templates)